[MOD/#68] 프로필 관리 뷰모델 분리
