+++ app/src/main/java/com/yello/presentation/main/profile/ProfileViewModel.kt
@@ -27,9 +27,6 @@ class ProfileViewModel @Inject constructor(
     private val _getListState = MutableLiveData<UiState<ProfileFriendsListModel?>>()
     val getListState: LiveData<UiState<ProfileFriendsListModel?>> = _getListState
 
-    private val _deleteUserState = MutableLiveData<UiState<Unit>>()
-    val deleteUserState: LiveData<UiState<Unit>> = _deleteUserState
-
     // TODO: 세터 설정 필요
     val _deleteFriendState = MutableLiveData<UiState<Unit>>()
     val deleteFriendState: LiveData<UiState<Unit>> = _deleteFriendState
@@ -95,21 +92,6 @@ class ProfileViewModel @Inject constructor(
                 _getListState.value = UiState.Success(it)
             }.onFailure {
                 _getListState.value = UiState.Failure(it.message.toString())
-            }
-        }
-    }
-
-    // 서버 통신 - 회원 탈퇴
-    fun deleteUserDataToServer() {
-        viewModelScope.launch {
-            _deleteUserState.value = UiState.Loading
-            runCatching {
-                profileRepository.deleteUserData()
-            }.onSuccess {
-                _deleteUserState.value = UiState.Success(it)
-                clearLocalInfo()
-            }.onFailure {
-                _deleteUserState.value = UiState.Failure(it.message.toString())
             }
         }
     }
