+++ app/src/main/java/com/yello/presentation/main/profile/manage/ProfileQuitDialog.kt
@@ -6,7 +6,7 @@ import android.content.res.Resources
 import android.os.Bundle
 import android.view.View
 import android.view.WindowManager
-import androidx.fragment.app.activityViewModels
+import androidx.fragment.app.viewModels
 import com.example.ui.base.BindingDialogFragment
 import com.example.ui.fragment.toast
 import com.example.ui.view.UiState
@@ -15,13 +15,12 @@ import com.kakao.sdk.user.UserApiClient
 import com.yello.R
 import com.yello.databinding.FragmentProfileQuitDialogBinding
 import com.yello.presentation.main.profile.ProfileViewModel
-import com.yello.util.context.yelloSnackbar
 import timber.log.Timber
 
 class ProfileQuitDialog :
     BindingDialogFragment<FragmentProfileQuitDialogBinding>(R.layout.fragment_profile_quit_dialog) {
 
-    private val viewModel by activityViewModels<ProfileViewModel>()
+    private val viewModel by viewModels<ProfileQuitViewModel>()
 
     override fun onStart() {
         super.onStart()
